requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.1
playsound==1.3.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pyobjc==10.3
pyobjc-core==10.3
pyobjc-framework-Accessibility==10.3
pyobjc-framework-Accounts==10.3
pyobjc-framework-AddressBook==10.3
pyobjc-framework-AdServices==10.3
pyobjc-framework-AdSupport==10.3
pyobjc-framework-AppleScriptKit==10.3
pyobjc-framework-AppleScriptObjC==10.3
pyobjc-framework-ApplicationServices==10.3
pyobjc-framework-AppTrackingTransparency==10.3
pyobjc-framework-AudioVideoBridging==10.3
pyobjc-framework-AuthenticationServices==10.3
pyobjc-framework-AutomaticAssessmentConfiguration==10.3
pyobjc-framework-Automator==10.3
pyobjc-framework-AVFoundation==10.3
pyobjc-framework-AVKit==10.3
pyobjc-framework-AVRouting==10.3
pyobjc-framework-BackgroundAssets==10.3
pyobjc-framework-BrowserEngineKit==10.3
pyobjc-framework-BusinessChat==10.3
pyobjc-framework-CalendarStore==10.3
pyobjc-framework-CallKit==10.3
pyobjc-framework-CFNetwork==10.3
pyobjc-framework-Cinematic==10.3
pyobjc-framework-ClassKit==10.3
pyobjc-framework-CloudKit==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-Collaboration==10.3
pyobjc-framework-ColorSync==10.3
pyobjc-framework-Contacts==10.3
pyobjc-framework-ContactsUI==10.3
pyobjc-framework-CoreAudio==10.3
pyobjc-framework-CoreAudioKit==10.3
pyobjc-framework-CoreBluetooth==10.3
pyobjc-framework-CoreData==10.3
pyobjc-framework-CoreHaptics==10.3
pyobjc-framework-CoreLocation==10.3
pyobjc-framework-CoreMedia==10.3
pyobjc-framework-CoreMediaIO==10.3
pyobjc-framework-CoreMIDI==10.3
pyobjc-framework-CoreML==10.3
pyobjc-framework-CoreMotion==10.3
pyobjc-framework-CoreServices==10.3
pyobjc-framework-CoreSpotlight==10.3
pyobjc-framework-CoreText==10.3
pyobjc-framework-CoreWLAN==10.3
pyobjc-framework-CryptoTokenKit==10.3
pyobjc-framework-DataDetection==10.3
pyobjc-framework-DeviceCheck==10.3
pyobjc-framework-DictionaryServices==10.3
pyobjc-framework-DiscRecording==10.3
pyobjc-framework-DiscRecordingUI==10.3
pyobjc-framework-DiskArbitration==10.3
pyobjc-framework-DVDPlayback==10.3
pyobjc-framework-EventKit==10.3
pyobjc-framework-ExceptionHandling==10.3
pyobjc-framework-ExecutionPolicy==10.3
pyobjc-framework-ExtensionKit==10.3
pyobjc-framework-ExternalAccessory==10.3
pyobjc-framework-FileProvider==10.3
pyobjc-framework-FileProviderUI==10.3
pyobjc-framework-FinderSync==10.3
pyobjc-framework-FSEvents==10.3
pyobjc-framework-GameCenter==10.3
pyobjc-framework-GameController==10.3
pyobjc-framework-GameKit==10.3
pyobjc-framework-GameplayKit==10.3
pyobjc-framework-HealthKit==10.3
pyobjc-framework-ImageCaptureCore==10.3
pyobjc-framework-InputMethodKit==10.3
pyobjc-framework-InstallerPlugins==10.3
pyobjc-framework-InstantMessage==10.3
pyobjc-framework-Intents==10.3
pyobjc-framework-IntentsUI==10.3
pyobjc-framework-IOBluetooth==10.3
pyobjc-framework-IOBluetoothUI==10.3
pyobjc-framework-IOSurface==10.3
pyobjc-framework-iTunesLibrary==10.3
pyobjc-framework-KernelManagement==10.3
pyobjc-framework-LatentSemanticMapping==10.3
pyobjc-framework-LaunchServices==10.3
pyobjc-framework-libdispatch==10.3
pyobjc-framework-libxpc==10.3
pyobjc-framework-LinkPresentation==10.3
pyobjc-framework-LocalAuthentication==10.3
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3
pyobjc-framework-MailKit==10.3
pyobjc-framework-MapKit==10.3
pyobjc-framework-MediaAccessibility==10.3
pyobjc-framework-MediaLibrary==10.3
pyobjc-framework-MediaPlayer==10.3
pyobjc-framework-MediaToolbox==10.3
pyobjc-framework-Metal==10.3
pyobjc-framework-MetalFX==10.3
pyobjc-framework-MetalKit==10.3
pyobjc-framework-MetalPerformanceShaders==10.3
pyobjc-framework-MetalPerformanceShadersGraph==10.3
pyobjc-framework-MetricKit==10.3
pyobjc-framework-MLCompute==10.3
pyobjc-framework-ModelIO==10.3
pyobjc-framework-MultipeerConnectivity==10.3
pyobjc-framework-NaturalLanguage==10.3
pyobjc-framework-NetFS==10.3
pyobjc-framework-Network==10.3
pyobjc-framework-NetworkExtension==10.3
pyobjc-framework-NotificationCenter==10.3
pyobjc-framework-OpenDirectory==10.3
pyobjc-framework-OSAKit==10.3
pyobjc-framework-OSLog==10.3
pyobjc-framework-PassKit==10.3
pyobjc-framework-PencilKit==10.3
pyobjc-framework-PHASE==10.3
pyobjc-framework-Photos==10.3
pyobjc-framework-PhotosUI==10.3
pyobjc-framework-PreferencePanes==10.3
pyobjc-framework-PushKit==10.3
pyobjc-framework-Quartz==10.3
pyobjc-framework-QuickLookThumbnailing==10.3
pyobjc-framework-ReplayKit==10.3
pyobjc-framework-SafariServices==10.3
pyobjc-framework-SafetyKit==10.3
pyobjc-framework-SceneKit==10.3
pyobjc-framework-ScreenCaptureKit==10.3
pyobjc-framework-ScreenSaver==10.3
pyobjc-framework-ScreenTime==10.3
pyobjc-framework-ScriptingBridge==10.3
pyobjc-framework-SearchKit==10.3
pyobjc-framework-Security==10.3
pyobjc-framework-SecurityFoundation==10.3
pyobjc-framework-SecurityInterface==10.3
pyobjc-framework-SensitiveContentAnalysis==10.3
pyobjc-framework-ServiceManagement==10.3
pyobjc-framework-SharedWithYou==10.3
pyobjc-framework-SharedWithYouCore==10.3
pyobjc-framework-ShazamKit==10.3
pyobjc-framework-Social==10.3
pyobjc-framework-SoundAnalysis==10.3
pyobjc-framework-Speech==10.3
pyobjc-framework-SpriteKit==10.3
pyobjc-framework-StoreKit==10.3
pyobjc-framework-Symbols==10.3
pyobjc-framework-SyncServices==10.3
pyobjc-framework-SystemConfiguration==10.3
pyobjc-framework-SystemExtensions==10.3
pyobjc-framework-ThreadNetwork==10.3
pyobjc-framework-UniformTypeIdentifiers==10.3
pyobjc-framework-UserNotifications==10.3
pyobjc-framework-UserNotificationsUI==10.3
pyobjc-framework-VideoSubscriberAccount==10.3
pyobjc-framework-VideoToolbox==10.3
pyobjc-framework-Virtualization==10.3
pyobjc-framework-Vision==10.3
pyobjc-framework-WebKit==10.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.14
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
QtAwesome==1.3.1
QtPy==2.4.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
